Move metadata_etl out of salesforce package
