+++ cumulusci/cumulusci.yml
@@ -9,19 +9,19 @@ tasks:
     add_page_layout_related_lists:
         group: Metadata Transformations
         description: Adds specified Related List to one or more Page Layouts.
-        class_path: cumulusci.tasks.salesforce.metadata_etl.AddRelatedLists
+        class_path: cumulusci.tasks.metadata_etl.AddRelatedLists
         options:
             namespace_inject: $project_config.project__package__namespace
     add_standard_value_set_entries:
         group: Metadata Transformations
         description: Adds specified picklist entries to a Standard Value Set.
-        class_path: cumulusci.tasks.salesforce.metadata_etl.AddValueSetEntries
+        class_path: cumulusci.tasks.metadata_etl.AddValueSetEntries
         options:
             namespace_inject: $project_config.project__package__namespace
     add_permission_set_perms:
         group: Metadata Transformations
         description: Adds specified Apex class access and Field-Level Security to a Permission Set.
-        class_path: cumulusci.tasks.salesforce.metadata_etl.AddPermissionSetPermissions
+        class_path: cumulusci.tasks.metadata_etl.AddPermissionSetPermissions
         options:
             namespace_inject: $project_config.project__package__namespace
     batch_apex_wait:
@@ -334,7 +334,7 @@ tasks:
     set_organization_wide_defaults:
         group: Metadata Transformations
         description: Sets the Organization-Wide Defaults for specific sObjects, and waits for sharing recalculation to complete.
-        class_path: cumulusci.tasks.salesforce.metadata_etl.SetOrgWideDefaults
+        class_path: cumulusci.tasks.metadata_etl.SetOrgWideDefaults
         options:
             namespace_inject: $project_config.project__package__namespace
     uninstall_managed:
